Updated to latest versions
This updates to the latest version of IntelliJ IDEA, Go and Markdown plugins.
Support for the .ignore plugin has been removed as it can be installed by the
user separately and the plugin itself can be resource heavy.

Readme has also been updated to contain extra docker parameters as reported
by users (fixes #2).